lambda/requirements.txt
pyyaml>=6.0
boto3>=1.28.0
requests>=2.31.0
